requirements.txt
asgiref==3.5.1
bitarray==2.7.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Django==4.0.4
django-axes==5.32.0
django-cors-headers==3.11.0
django-ipware==4.0.2
django-rest-durin==1.0.0
djangorestframework==3.13.1
humanize==4.1.0
idna==3.3
pdfreader==0.1.12
Pillow==9.5.0
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodome==3.17
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
